Create VirtualhostFuzzer
This script curls a target IP and domain, its useful after running a tool like ffuf to view web page code. I created it for a virtualhosts challenge on hackthebox. 

To use the tool:

Change the IP and the domain at the bottom.
Create a wordlist of virtualhosts you've grabbed using ffuf.
Then run the tool like this:
./fuzzer.sh vhosts.txt